Fixed the bug for parallel RBS, was not a rac condition but two flaws in the stop logic
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@13467 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/search.hh
@@ -379,7 +379,7 @@ namespace Gecode { namespace Search {
     /// Return true if meta engine must be stopped
     virtual bool stop(const Statistics& s, const Options& o);
     /// Set current limit for the engine to \a l fails
-    void limit(const Search::Statistics& s, unsigned int l);
+    void limit(const Search::Statistics& s, unsigned long int l);
     /// Return the stop object to control the engine
     Stop* enginestop(void) const;
     /// Return whether the engine has been stopped
